tools/requirements.txt
pytest==8.3.3
pytest-cov==5.0.0
pylint==3.2.6
pyyaml==6.0.2
python-frontmatter==1.1.0